+++ app/recorder/renderer.js
@@ -30,13 +30,21 @@ function notify( txt ) {
   setTimeout( () => notifyEl.classList.toggle( "has-go", false ), 1000 );
 }
 
-function loadUrl( url ) {
-  localStorage.setItem( STORAGE_URL, url );
-  webview.src = url;
+function onceWebviewReady() {
+  webview.removeEventListener( "dom-ready", onceWebviewReady );
   webview.send( "goto" );
   info.classList.add( "is-hidden" );
   webview.classList.remove( "is-hidden" );
   okBtn.removeAttribute( "disabled" );
+}
+
+function loadUrl( url ) {
+  localStorage.setItem( STORAGE_URL, url );
+  webview.src = url;
+
+  webview.addEventListener( "dom-ready", onceWebviewReady );
+
+
   commands.push({
     target: "page",
     method: "goto",